Extract `deep-equal` tests from `sinon-test`.
Pulls out a suite into `test/util/core/deep-equal-test` for better organisation.
